feat(Waiting): modify peopleCount data type Long -> Integer

+++ src/main/java/com/matdang/seatdang/waiting/controller/WaitingController.java
@@ -160,7 +160,7 @@ public class WaitingController {
                                 .waitingNumber(i)
                                 .waitingOrder(i)
                                 .storeId(1L)
-                                .customerInfo(new CustomerInfo(i, "010-1111-1111", ((long) (Math.random() * 3 + 1))))
+                                .customerInfo(new CustomerInfo(i, "010-1111-1111", ((int) (Math.random() * 3 + 1))))
                                 .waitingStatus(value)
                                 .visitedTime(null)
                                 .build());
@@ -174,7 +174,7 @@ public class WaitingController {
                     .waitingNumber(i)
                     .waitingOrder(i)
                     .storeId(2L)
-                    .customerInfo(new CustomerInfo(i, "010-1111-1111", ((long) (Math.random() * 3 + 1))))
+                    .customerInfo(new CustomerInfo(i, "010-1111-1111", ((int) (Math.random() * 3 + 1))))
                     .waitingStatus(WaitingStatus.WAITING)
                     .visitedTime(null)
                     .build());
